fix: Code style fix and refactorings
